backend/pyproject.toml
[tool.poetry]
name = "fianlbackend"
version = "0.1.0"
description = "Backend environment for Final project"
authors = ["hjhhijk"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
multipart = "^0.2.4"
typing = "3.5"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
jinja2 = "^3.1.2"
sqlalchemy = "^2.0.18"
fastapi-sqlalchemy = "^0.2.1"
pymysql = "^1.1.0"
mysqlclient = "2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"